Fix various Message thingsjj
Change out title() with full_name and name.
Add joined loading for some relations.
Fix so that viewing a message ID is possible both for sent and received.
Fix receivers so that it shows soemthing.
Add a short content to the table.
Add Menu
Change from list to set() in form submit code.

Change-Id: I56869a8ae9e2536d4a2a5dda2565cfd30151f481